Move maker-specific actors from daemon to maker
