+++ daemon-tests/Cargo.toml
@@ -4,8 +4,6 @@ version = "0.1.0"
 edition = "2021"
 publish = false
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 anyhow = "1"
 async-trait = "0.1"
@@ -17,6 +15,7 @@ model = { path = "../model" }
 rand = "0.6"
 rust_decimal = "1.23"
 rust_decimal_macros = "1.23"
+maker = { path = "../maker" }
 time = "0.3"
 tokio = { version = "1", features = ["rt-multi-thread", "macros", "sync", "net"] }
 tokio-tasks = { path = "../tokio-tasks", features = ["xtra"] }
